L8 release 1

+++ L8/src/main/java/com/otus/alexeenko/l8/services/DataBaseService.java
@@ -2,12 +2,10 @@ package com.otus.alexeenko.l8.services;
 
 import com.otus.alexeenko.l8.services.datasets.BaseDataSet;
 
-import java.math.BigInteger;
-
 /**
  * Created by Vsevolod on 11/06/2017.
  */
 public interface DataBaseService {
     <T extends BaseDataSet> void save(T dataSet);
-    <T extends BaseDataSet> T load(BigInteger id, Class<T> clazz);
+    <T extends BaseDataSet> T load(long id, Class<T> clazz);
 }
